Adding working, modular popouts and making more MVC.
DetectorContainer replaces a lot of functionality of what is now called
DetectorPlotView (used to be DetectorPanel). DetectorContainer handles
popping in and out views, and delegates Formatter interface to owned
widgets. DetectorPopout is just a tab view for full data in
DetectorContainer. DetectorPlotView is now a sibling of
DetectorTableView, DetectorParametersView, and DetectorCommandView (but
those are unimplemented).